refactor: rename useEvent.ts file to use-event.ts
