+++ app/routes/team+/$team.$event+/reviews+/__components/proposals-list-page/filters/sort-menu.tsx
@@ -6,7 +6,7 @@ import { Fragment } from 'react';
 
 import { button } from '~/design-system/buttons.tsx';
 import { MenuTransition } from '~/design-system/transitions.tsx';
-import { useEvent } from '~/routes/team+/$team.$event+/__components/useEvent.tsx';
+import { useEvent } from '~/routes/team+/$team.$event+/__components/use-event.tsx';
 
 import { sortByDatesOptions, sortByReviewsOptions } from './filters.ts';
