+++ app/routes/team+/$team.$event+/settings+/tracks.tsx
@@ -12,7 +12,7 @@ import { H2, Subtitle } from '~/design-system/typography.tsx';
 import { requireSession } from '~/libs/auth/session.ts';
 import { toast } from '~/libs/toasts/toast.server.ts';
 
-import { useEvent } from '../__components/useEvent.tsx';
+import { useEvent } from '../__components/use-event.tsx';
 import { NewTrackButton } from './__components/save-track-form.tsx';
 import { TrackList } from './__components/track-list.tsx';
